Extract 抽出两个 Adapter
