+++ app/src/main/java/com/tantuo/didicar/fragment/ContentFragment.java
@@ -1,13 +1,12 @@
 package com.tantuo.didicar.fragment;
 
-import android.support.v4.view.PagerAdapter;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.RadioGroup;
 
 import com.jeremyfeinstein.slidingmenu.lib.SlidingMenu;
 import com.tantuo.didicar.MainActivity;
 import com.tantuo.didicar.R;
+import com.tantuo.didicar.adapter.ContentFragmentAdapter;
 import com.tantuo.didicar.base.BaseMenuFragment;
 import com.tantuo.didicar.base.BasePager;
 import com.tantuo.didicar.pager.CardServicePager;
@@ -85,40 +84,43 @@ public class ContentFragment extends BaseMenuFragment {
 
 
 
+
     }
 
+    //实现点击屏幕下方Radiogroup切换不同ViewPager最重要的方法
     class MyOnCheckedChangeListener implements RadioGroup.OnCheckedChangeListener {
 
         @Override
         public void onCheckedChanged(RadioGroup group, int checkedId) {
-            switch (checkedId){
+            switch (checkedId) {
                 case R.id.rb_home:
-                    viewpager.setCurrentItem(0,true);
+                    viewpager.setCurrentItem(0, true);
                     //smoothScroll代表是否滑动渐入或者渐出 by:Tantuo
                     isEnableSlidingMenu(SlidingMenu.TOUCHMODE_NONE);
                     break;
                 case R.id.rb_map:
-                    viewpager.setCurrentItem(1,true);
+                    viewpager.setCurrentItem(1, true);
                     isEnableSlidingMenu(SlidingMenu.TOUCHMODE_NONE);
                     break;
                 case R.id.rb_cardservice:
-                    viewpager.setCurrentItem(2,true);
+                    viewpager.setCurrentItem(2, true);
                     isEnableSlidingMenu(SlidingMenu.TOUCHMODE_NONE);
                     break;
                 case R.id.rb_govaffair:
-                    viewpager.setCurrentItem(3,true);
+                    viewpager.setCurrentItem(3, true);
                     isEnableSlidingMenu(SlidingMenu.TOUCHMODE_NONE);
                     break;
                 case R.id.rb_setting:
-                    viewpager.setCurrentItem(4,true);
+                    viewpager.setCurrentItem(4, true);
                     isEnableSlidingMenu(SlidingMenu.TOUCHMODE_NONE);
                     break;
             }
 
         }
     }
+
     /**
-     根据传入的参数设置是否让SlidingMenu可以滑动
+     * 根据传入的参数设置是否让SlidingMenu可以滑动
      */
     private void isEnableSlidingMenu(int touchmodeFullscreen) {
         MainActivity mainActivity = (MainActivity) context;
@@ -126,43 +128,4 @@ public class ContentFragment extends BaseMenuFragment {
     }
 
 
-    public class ContentFragmentAdapter extends PagerAdapter {
-
-        private final ArrayList<BasePager> basePagers;
-
-        public ContentFragmentAdapter(ArrayList<BasePager> basePagers) {
-            this.basePagers = basePagers;
-        }
-
-        @Override
-        public int getCount() {
-
-            return basePagers.size();
-        }
-
-        @Override
-        public Object instantiateItem(ViewGroup container, int position) {
-            BasePager basePager = basePagers.get(position);//各个页面的实例
-            View rootView = basePager.rootView;//各个子页面
-
-            //调用initData()
-            //注意：这里的initData()看上去调用的是basePager,肯定是调用的是各个子类pager的。by:Tantuo
-            basePager.initData();//初始化数据
-
-            container.addView(rootView);
-            return rootView;
-        }
-
-
-        @Override
-        public boolean isViewFromObject(View view, Object object) {
-            return view == object;
-        }
-
-
-        @Override
-        public void destroyItem(ViewGroup container, int position, Object object) {
-            container.removeView((View) object);
-        }
-    }
 }
